Functional version 2
Send and recive prices and productions from Generators and Controllers
to the Central eNode. Also can calculate the resulting frequency of the
grid and and control the problem with Stability Controllers. WT, PV,
Hydro, Diesel and Centra eNode implemented in the Grid. Consumption of
the controllers done, but not the fuel/water level.